refactor: refactoring dashBoard  page
