+++ src/types/dashboardTypes.ts
@@ -1,0 +1,10 @@
+export type DashBoardArticle = {
+  id: string,
+  name: string
+}
+
+export type DashBoardPerson = {
+  email: string;
+  name: string;
+  status: boolean
+}
